requirements.txt
asgiref==3.9.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.4
fhirclient==4.3.1
idna==3.10
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
